Moving checkerboard out of boardgame/ui
