+++ examples/modules/chess/components/board.js
@@ -9,7 +9,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Chess from 'chess.js';
-import Checkerboard from '../../../../src/ui/checkerboard';
+import Checkerboard from './checkerboard';
 import Token from '../../../../src/ui/token';
 import Bishop from './pieces/bishop';
 import King from './pieces/king';
