+++ examples/modules/chess/components/checkerboard.js
